requirements.txt
Flask==2.0.1
python-dotenv==0.18.0
stripe==2.60.0